pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "llm-usage-tracker"
version = "0.1.0"
description = "Tracks usage limits across Claude, ChatGPT, and Gemini subscriptions"
requires-python = ">=3.12"
dependencies = [
    "playwright>=1.44.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",
    "typer>=0.12.0",
    "rich>=13.7.0",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "apscheduler>=3.10.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "greenlet>=3.0.0",
]

[project.scripts]
llm-tracker = "backend.cli:app"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
asyncio_mode = "auto"
